Finalise Update 1.5.0

+++ modules/configManager.py
@@ -3,7 +3,7 @@ from time import sleep
 from termcolor import colored
 import os
 
-def_config_version = 1.4
+def_config_version = 1.5
 
 class Config_Manager():
     
@@ -13,8 +13,13 @@ class Config_Manager():
             "proxies": True,
             "checkForUpdates": True,
             "oneTimeDownload": True,
+            "advancedMode": False,
             "user_credentials": {
                 "e621": {
+                    "apiUser": "",
+                    "apiKey": ""
+                },
+                "e6ai": {
                     "apiUser": "",
                     "apiKey": ""
                 },
@@ -24,10 +29,7 @@ class Config_Manager():
                 },
                 "furbooru": {
                     "apiKey": ""
-                },
-                "github": {
-                    "apiKey": ""
-                },
+                }
             },
             "blacklisted_tags": [
                 "example1",
@@ -46,10 +48,18 @@ class Config_Manager():
     def reader():
         if os.path.exists("config.json"):
             with open("config.json", "r") as cf:
-                config = json.load(cf)
-                config_version = config["version"]
-                
-            if config_version < def_config_version:
+                try:
+                    config = json.load(cf)
+                    config_version = config["version"]
+                    advanced_mode = config["advancedMode"]
+                except:
+                    config_version = 0
+                    advanced_mode = False
+            
+            if advanced_mode == True:
+                return config
+            
+            elif config_version < def_config_version and advanced_mode != True:
                 print(colored("You are using an outdated config version! Old one is backed up. Please reconfigure the new one.", "green"))
                 if os.path.exists("old_config.json"):
                     os.remove("old_config.json")
@@ -63,12 +73,4 @@ class Config_Manager():
         else:
             return 0
         # 0 means unsuccessful
-         
-    def compare():
-        # TODO: Add working config compare method instead of simple config version number
-        print("how in the hell did you even manage to get this printed?")
-                        
-    def updater():
-        # TODO: Add working config updater
-        print("how in the hell did you even manage to get this printed?")
